refactor(agent): split Host mappings by direction
`mapping.ts` held both directions of the protocol/Runtime adaptation. The
write side — `toRuntimeHistory`, `toRuntimeInputParts` and the attachment
map they consume — is turn material assembly and moves to
`turnRuntimeInput.ts` next to the other turn stages. The read side keeps
the file (renamed `runtimeProjection.ts`) and now states which direction
it owns. Tests split along the same seam.
